flask_app/requirements.txt
flask==3.1
opencv-python
kano-cv==3.3
pycocotools
ultralytics
pyyaml
pillow